+++ js/leaderboard.js
@@ -18,8 +18,8 @@ if(q != undefined){
 
 // Check for parameters
 if (params == 0) {
-    sqlString = "select A,B,C,D,E order by C desc";
-    loadResults(sqlString, mySpreadsheet);
+    sqlQuery = "select A,B,C,D,E order by C desc";
+    loadResults(sqlQuery, mySpreadsheet);
 }
 
 // define function to load results