+++ js/postponed.js
@@ -1,9 +1,9 @@
 var mySpreadsheet = 'https://docs.google.com/spreadsheets/d/1xqGTbkgosPqSRCkZ6xKj1c01sRRZkg0qeNeN2hrkFSI/pubhtml?gid=1867154294';
 var bandsTemplate = Handlebars.compile($('#bands-template').html());
 
-sqlString = "select A,B,I where I contains 'Postponed' order by A desc";
+sqlQuery = "select A,B,I where I contains 'Postponed' order by A desc";
 
-loadResults(sqlString);
+loadResults(sqlQuery);
 function loadResults(sql){
   $('#bands').sheetrock({
     url: mySpreadsheet,