+++ sbdb-1.0/concepts.js
@@ -1,9 +1,9 @@
 var mySpreadsheet = 'https://docs.google.com/spreadsheets/d/1xqGTbkgosPqSRCkZ6xKj1c01sRRZkg0qeNeN2hrkFSI/pubhtml?gid=1847002595';
 var bandsTemplate = Handlebars.compile($('#bands-template').html());
 
-sqlString = "select O, count(C), max(A), min(B) group by O order by O";
+sqlQuery = "select O, count(C), max(A), min(B) group by O order by O";
 
-loadResults(sqlString);
+loadResults(sqlQuery);
 function loadResults(sql){
   $('#bands').sheetrock({
     url: mySpreadsheet,