+++ sbdb-1.0/video.js
@@ -91,9 +91,9 @@ setTimeout(function() {
   message = '<iframe class="embed-responsive-item" src="' + video + '"></iframe>';
   document.getElementById('video').innerHTML = message;
   document.getElementById('searchTerm').innerHTML = "<h2>" + year + " " + band + " String Band</h2>";
-  sqlString = "select A,B,C,D,E,F,M,L,V,W,G,H,I,J,K,X,Q,R,S,T where A = " + year + " order by A desc";
+  sqlQuery = "select A,B,C,D,E,F,M,L,V,W,G,H,I,J,K,X,Q,R,S,T where A = " + year + " order by A desc";
   document.getElementById('results-tag').innerHTML = year + " Results";
-  loadResults(sqlString, '#bands');
+  loadResults(sqlQuery, '#bands');
 
 }, delay);