+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 12/22/2023
 - Use decodeURIComponent to decode search paramaters back to original form, fixing search queries for double quotes.
+- Rename "sqlString" variable to "sqlQuery" for clarity.
+- Update parade live stream link.
 
 11/28/2023
 - Remove links to V1 of Database and add message that V1 is no longer maintained.